perf: avoid reallocation by reusing the available slack space
use extra 4 bytes of the allocation to track the extra capacity (slack
space). The entire allocation is split in two and can have more space
than it is required to serialize an account, thus when account grows its
data field the extra space can be used for growth instead of forcing
reallocation.